Um monte de coisa foi adicionada

+++ src (copy)/ep1/fitness/Fitness.java
@@ -4,7 +4,7 @@ import ep1.Cromossomo;
 
 public interface Fitness {
     
-    public double calcularFitness(Cromossomo individuo);
+    public int calcularFitness(Cromossomo individuo);
     
     public boolean individuoApto(Cromossomo individuo);
